Removed the requirement that subchannels needed to have an oscillation pattern, it now defaults to no oscillations if absent

+++ examples/make_friend_example_macro.c
@@ -1,7 +1,6 @@
 void make_friend_example_macro(){
 
-
-
+    
 //Lets say you have a file, 1e1p.root containing a TTree called events. 
 TFile *fin = new TFile("1e1p.root","read");
 TTree *tin = (TTree*)fin->Get("events");
@@ -9,7 +8,6 @@ TTree *tin = (TTree*)fin->Get("events");
 //We want to reweight this by some function of a branch, say ereco
 double ereco;
 tin->SetBranchAddress("ereco",&ereco);
-
 
 //We want to store the new weights in another file and TTtee that we can easily assign as a FriendTree to the original
 TFile *fout = new TFile("1e1p_weights.root","recreate");
@@ -20,9 +18,8 @@ TTree *tout = new TTree("signalweights","signalweights");
 double wei;
 tout->Branch("leeweight", &wei,"leeweight/D");
 
-
 //Now we loop over all events in the input file, and calculate a weight based on our branch 
-for(int i=0; i< tin->GetEntries(); ++i){
+for(size_t i=0; i< tin->GetEntries(); ++i){
     tin->GetEntry(i);
 
     //This is just a nonsence quick test, you would put any weights appropiate here
@@ -35,18 +32,14 @@ for(int i=0; i< tin->GetEntries(); ++i){
         wei = 4.0;
     }
 
-    //and fill the new tree. Because its being filled once for every event in the input TTree in unison, we know the entries match up and we know it will work well as a TTree.
+    //and fill the new tree. Because its being filled once for every event in the input TTree in unison, we know the entries match up and thus it be correctly assigned as a FriendTree.
     tout->Fill();
 }
-
 tout->Write();
 fout->Close();
-
-
 //The weights can then be easily accessed by assigning signalweights as a FriendTree via
 // tin->AddFriend("signalweights","1e1p_weights.root");
 // tin->Draw("ereco", "signalweights.leeweight");
-
 
 return;
 }
